backend/requirements.txt
autoflake==2.3.1
black==25.1.0
paho-mqtt==2.1.0
pyclean==3.1.0
pytest-cov==7.0.0
pytest==8.4.2
python-dotenv==1.1.1
